commit initial octopart provider (draft)
When I start to work to this project I was focused mainly on the task of reverse engineering the digikey barcode and than use it to retrieve info on web.
Unfortunately only few digikey components are indexed by search engine and digikey website has a lot of anti-bot check in place. Could the octopart provider could fix those problems?

You need a OCTOPART token here see:
https://octopart.com/api/docs/v3/overview